+++ deployments/chef/spec/unit/recipes/default_spec.rb
@@ -100,7 +100,7 @@ describe 'splunk_otel_collector::default' do
             'splunk_access_token' => 'test123',
             'splunk_realm' => 'test',
             'with_auto_instrumentation' => true,
-            'auto_instrumentation_version' => '0.158.0',
+            'auto_instrumentation_version' => '0.159.0',
             'with_auto_instrumentation_sdks' => %w(java),
           }
         end.converge described_recipe